Configured Nginx reverse proxy
- Configured Nginx as reverse proxy to serve API requests
- Fixed intermittent 403 errors by using 127.0.0.1 instead of localhost

+++ config/stop.bash
@@ -2,5 +2,14 @@
 # https://www.nginx.com/resources/wiki/start/topics/tutorials/commandline/
 # http://stackoverflow.com/questions/59895/getting-the-source-directory-of-a-bash-script-from-within
 dir="$(cd "$(dirname "$0")" && cd .. && pwd)"
+
+# Stop NGINX if nginx.pid file exists
+if [ -f "$dir/logs/nginx.pid" ]; then
+  echo "Stopping NGINX server..."
+
 # https://www.nginx.com/resources/wiki/start/topics/tutorials/commandline/
-kill -QUIT "$(< "$dir/logs/nginx.pid")"
+  kill -QUIT "$(< "$dir/logs/nginx.pid")"  # Stop NGINX gracefully by sending QUIT signal
+  echo "NGINX stopped successfully."
+else
+  echo "No NGINX process found to stop."
+fi
